It was lonely for a day or so until one morning some man, more recently arrived than I, stopped me on the road.
"How do you get to West Egg village?" he asked helplessly.

I told him. And as I walked on I was lonely no longer.
I was a guide, a pathfinder, an original settler.
He had casually conferred on me the freedom of the neighborhood.

And so with the sunshine and the great bursts of leaves growing on
the trees—just as things grow in fast movies—I had that familiar conviction
that life was beginning over again with the summer.